chore(AssetStorage): move loader to dedicated sub file
