+++ libgraphics/source/AssetsStorage/Loaders/AssetsStorage_gltf.cxx
@@ -1,0 +1,14 @@
+#include "pivot/graphics/AssetStorage.hxx"
+
+#include "pivot/graphics/DebugMacros.hxx"
+
+namespace pivot::graphics
+{
+
+bool AssetStorage::loadGltfModel(const std::filesystem::path &path)
+{
+    logger.err("AssetStorage/GLTF") << "Not implemented";
+    return false;
+}
+
+}    // namespace pivot::graphics
