+++ libgraphics/source/AssetsStorage/Loaders/AssetsStorage_ktx.cxx
@@ -1,0 +1,14 @@
+#include "pivot/graphics/AssetStorage.hxx"
+
+#include "pivot/graphics/DebugMacros.hxx"
+
+namespace pivot::graphics
+{
+
+bool AssetStorage::loadKtxTexture(const std::filesystem::path &path)
+{
+    logger.err("AssetStorage/KTX") << "Not implemented";
+    return false;
+}
+
+}    // namespace pivot::graphics
